% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flightist.R
\docType{data}
\name{Flightist}
\alias{Flightist}
\title{Flight Prices}
\source{
\url{https://www.kaggle.com/datasets/ibrahimelsayed182/plane-ticket-price}
}
\usage{
data(Flightist)
}
\arguments{
\item{Airline}{\itemize{
\item Names of airlines
\item They are Air Asia, Air India, GoAir, IndiGo, Jet Airways, Jet Airways Business, Multiple carriers, Multiple carriers Premium economy, SpiceJet, Trujet, Vistara, and Vistara Premium economy.
}}

\item{Date_of_Journey}{\itemize{
\item Flight date
\item Dates found in the dataset: 03/01/2019, 03/03/2019, 03/06/2019, 03/09/2019, 03/12/2019, 03/15/2019, 03/18/2019, 03/21/2019, 03/24/2019, 03/27/2019, 04/01/2019, 04/03/2019, 04/06/2019, 04/09/2019, 04/12/2019, 04/15/2019, 04/18/2019, 04/21/2019, 04/24/2019, 04/27/2019, 05/01/2019, 05/03/2019, 05/06/2019, 05/09/2019, 05/12/2019, 05/15/2019, 05/18/2019, 05/21/2019, 05/24/2019, 05/27/2019, 06/01/2019, 06/03/2019, 06/06/2019, 06/09/2019, 06/12/2019, 06/15/2019, 06/18/2019, 06/21/2019, 06/24/2019, 06/27/2019
}}

\item{Source}{\itemize{
\item Origin airport
\item Airport locations: Banglore, Chennai, Delhi, Kolkata, Mumbai
}}

\item{Destination}{\itemize{
\item Destination
\item Airport locations: Banglore, Cochin, Delhi, Hyderabad, Kolkata, New Delhi
}}

\item{Route}{Specifies the route the aircraft will follow to reach its destination from its origin, including stops en route.}

\item{Dep_Time}{Departure time (01:00 PM - 12:55 PM)}

\item{Arrival_Time}{Arrival time (01:00 AM - 12:55 PM)}

\item{Duration}{The time it takes for the flight to fly from source to destination.}

\item{Total_Stops}{\itemize{
\item Total stops
\item Number of stops: 1 stop, 2 stops, 3 stops, 4 stops, non-stop
}}

\item{Additional_Info}{\itemize{
\item Additional information about the flight
\item They are 1 Long layover, 1 Short layover, 2 Long layover, Business class, Change airports, In-flight meal not included, No check-in baggage included, No info, No Info, and Red-eye flight.
}}

\item{Price}{Flight price}
}
\description{
This dataset contains 10683 observations and 11 variables. These variables include prices and other attributes of flights.
}
\keyword{datasets}
