% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get a One-Way OD Ticket.R
\name{One_Way}
\alias{One_Way}
\title{Buy One-Way Tickets}
\usage{
One_Way(data, SRC, DST, date, PAX)
}
\arguments{
\item{data}{the dataset (use "Flightist.csv" or find the relevant dataset)}

\item{SRC}{Origin airport}

\item{DST}{Destination airport}

\item{date}{Date of flight}

\item{PAX}{Number of passengers}
}
\value{
Identify the data type in a console and view data in a spreadsheet-style data viewer.
}
\description{
Input the inputs and output a dataframe that shows the trips matching the given inputs starting at the lowest prices and finishing at the highest prices.
}
\examples{
One_Way("Flightist.csv", "Banglore", "New Delhi", "03/03/2019", 2)
One_Way("Flightist.csv", "Delhi", "Cochin", "06/09/2019", 120)
}
