% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get a One-Way Airline OD Ticket.R
\name{Preferable_Airline_One_Way}
\alias{Preferable_Airline_One_Way}
\title{Booking One-Way Tickets with a Preferred Airline}
\usage{
Preferable_Airline_One_Way(data, SRC, DST, date, PAX, airline)
}
\arguments{
\item{data}{the dataset (use "Flightist.csv" or find the relevant dataset)}

\item{SRC}{Origin airport}

\item{DST}{Destination airport}

\item{date}{Date of flight}

\item{PAX}{Number of passengers}

\item{airline}{Preferred airline}
}
\value{
Identify the data type in a console and view data in a spreadsheet-style data viewer.
}
\description{
Input the inputs and output a dataframe that shows the trips matching the given inputs starting at the lowest prices and finishing at the highest prices.
}
\examples{
Preferable_Airline_One_Way("Flightist.csv", "Banglore", "New Delhi", "03/03/2019", 2, "Vistara Premium economy")
Preferable_Airline_One_Way("Flightist.csv", "Delhi", "Cochin", "06/09/2019", 120, "IndiGo")
}
