% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get OD Roundtrip Tickets.R
\name{Roundtrip}
\alias{Roundtrip}
\title{Purchase Roundtrip Tickets}
\usage{
Roundtrip(data, SRC, DST, date, PAX, SRC2, DST2, dateReturn, PAX2)
}
\arguments{
\item{data}{the dataset (use "Flightist.csv" or find the relevant dataset)}

\item{SRC}{Origin airport}

\item{DST}{Destination airport}

\item{date}{Date of flight}

\item{PAX}{Number of passengers}

\item{SRC2}{Departures can be made either from the same airport as the destination or from another airport.}

\item{DST2}{The airport where you will be taken back to your origin location}

\item{dateReturn}{Return date}

\item{PAX2}{Number of passengers}
}
\value{
Identify the data types in a console and view the data in spreadsheet-like views.
}
\description{
Input the inputs and output a dataframe that shows the departure and arrival schedules that match the given inputs.
}
\examples{
Roundtrip("Flightist.csv", "Banglore", "New Delhi", "03/03/2019", 3, "Kolkata", "Banglore", "03/24/2019", 3)
Roundtrip("Flightist.csv", "Banglore", "New Delhi", "03/01/2019", 2, "Delhi", "Cochin", "03/21/2019", 2)
}
